Modules/requirements.txt
streamlit
openai
langchain
tiktoken
pandas
#sqlite3
requests
langchain_groq
langgraph
httpx
python-dotenv
win32com
langchain_core
docx
PyPDF2
crewai